chore: refactor commit stuff into it's own file
