+++ internal/git/git_test.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	"github.com/ConfigButler/gitops-reverser/internal/ssh"
 	"github.com/ConfigButler/gitops-reverser/internal/types"
@@ -194,167 +193,6 @@ func TestToGitPath_SpecialCharacters(t *testing.T) {
 	}
 }
 
-func TestGetCommitMessage_CreateOperation(t *testing.T) {
-	obj := &unstructured.Unstructured{}
-	obj.SetName("test-pod")
-	obj.SetNamespace("default")
-	obj.SetKind("Pod")
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "pods",
-			Namespace: "default",
-			Name:      "test-pod",
-		},
-		Operation: "CREATE",
-		UserInfo: UserInfo{
-			Username: "john.doe@example.com",
-		},
-		Path: "",
-	}
-
-	message := GetCommitMessage(event)
-	expected := "[CREATE] v1/pods/test-pod"
-	assert.Equal(t, expected, message)
-}
-
-func TestGetCommitMessage_UpdateOperation(t *testing.T) {
-	obj := &unstructured.Unstructured{}
-	obj.SetName("my-service")
-	obj.SetNamespace("production")
-	obj.SetKind("Service")
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "services",
-			Namespace: "production",
-			Name:      "my-service",
-		},
-		Operation: "UPDATE",
-		UserInfo: UserInfo{
-			Username: "system:serviceaccount:kube-system:deployment-controller",
-		},
-		Path: "prod-repo",
-	}
-
-	message := GetCommitMessage(event)
-	expected := "[UPDATE] v1/services/my-service"
-	assert.Equal(t, expected, message)
-}
-
-func TestGetCommitMessage_DeleteOperation(t *testing.T) {
-	obj := &unstructured.Unstructured{}
-	obj.SetName("old-config")
-	obj.SetNamespace("staging")
-	obj.SetKind("ConfigMap")
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "configmaps",
-			Namespace: "staging",
-			Name:      "old-config",
-		},
-		Operation: "DELETE",
-		UserInfo: UserInfo{
-			Username: "admin",
-		},
-		Path: "staging-repo",
-	}
-
-	message := GetCommitMessage(event)
-	expected := "[DELETE] v1/configmaps/old-config"
-	assert.Equal(t, expected, message)
-}
-
-func TestGetCommitMessage_ClusterScopedResource(t *testing.T) {
-	obj := &unstructured.Unstructured{}
-	obj.SetName("my-namespace")
-	obj.SetKind("Namespace")
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "namespaces",
-			Namespace: "",
-			Name:      "my-namespace",
-		},
-		Operation: "CREATE",
-		UserInfo: UserInfo{
-			Username: "cluster-admin",
-		},
-		Path: "cluster-repo",
-	}
-
-	message := GetCommitMessage(event)
-	expected := "[CREATE] v1/namespaces/my-namespace"
-	assert.Equal(t, expected, message)
-}
-
-func TestGetCommitMessage_EmptyUsername(t *testing.T) {
-	obj := &unstructured.Unstructured{}
-	obj.SetName("test-pod")
-	obj.SetNamespace("default")
-	obj.SetKind("Pod")
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "pods",
-			Namespace: "default",
-			Name:      "test-pod",
-		},
-		Operation: "CREATE",
-		UserInfo: UserInfo{
-			Username: "", // Empty username
-		},
-		Path: "",
-	}
-
-	message := GetCommitMessage(event)
-	expected := "[CREATE] v1/pods/test-pod"
-	assert.Equal(t, expected, message)
-}
-
-func TestGetCommitMessage_SpecialCharactersInNames(t *testing.T) {
-	obj := &unstructured.Unstructured{}
-	obj.SetName("test-pod.with.dots")
-	obj.SetNamespace("test-ns_with_underscores")
-	obj.SetKind("Pod")
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "pods",
-			Namespace: "test-ns_with_underscores",
-			Name:      "test-pod.with.dots",
-		},
-		Operation: "UPDATE",
-		UserInfo: UserInfo{
-			Username: "user@domain.com",
-		},
-		Path: "",
-	}
-
-	message := GetCommitMessage(event)
-	expected := "[UPDATE] v1/pods/test-pod.with.dots"
-	assert.Equal(t, expected, message)
-}
-
 func TestGetAuthMethod_ValidKey(t *testing.T) {
 	// Since we're testing with a fake key, we expect this to fail
 	// In a real scenario, this would be a valid SSH private key
@@ -445,85 +283,12 @@ func TestCommitFile_Structure(t *testing.T) {
 	assert.Contains(t, string(file.Content), "name: test-pod")
 }
 
-func TestIntegration_FilePathAndCommitMessage(t *testing.T) {
-	// Test that file path and commit message work together correctly
-	obj := &unstructured.Unstructured{}
-	obj.SetName("integration-test-pod")
-	obj.SetNamespace("integration-test")
-	obj.SetKind("Pod")
-
-	identifier := types.ResourceIdentifier{
-		Group:     "",
-		Version:   "v1",
-		Resource:  "pods",
-		Namespace: "integration-test",
-		Name:      "integration-test-pod",
-	}
-
-	event := Event{
-		Object:     obj,
-		Identifier: identifier,
-		Operation:  "CREATE",
-		UserInfo: UserInfo{
-			Username: "integration-test-user",
-		},
-		Path: "integration-repo",
-	}
-
-	filePath := identifier.ToGitPath()
-	commitMessage := GetCommitMessage(event)
-
-	expectedPath := "v1/pods/integration-test/integration-test-pod.yaml"
-	expectedMessage := "[CREATE] v1/pods/integration-test-pod"
-
-	assert.Equal(t, expectedPath, filePath)
-	assert.Equal(t, expectedMessage, commitMessage)
-
-	// Verify they reference the same resource
-	assert.Contains(t, filePath, "integration-test-pod")
-	assert.Contains(t, commitMessage, "integration-test-pod")
-	assert.Contains(t, filePath, "integration-test")
-}
-
 func TestEdgeCases_NilIdentifier(t *testing.T) {
 	// Test behavior with empty identifier
 	identifier := types.ResourceIdentifier{}
 	path := identifier.ToGitPath()
 	// Empty identifier will produce minimal path
 	assert.NotEmpty(t, path)
-}
-
-func TestCommitMessage_AllOperations(t *testing.T) {
-	obj := &unstructured.Unstructured{}
-	obj.SetName("test-resource")
-	obj.SetNamespace("test-ns")
-	obj.SetKind("TestKind")
-
-	operations := []string{"CREATE", "UPDATE", "DELETE", "CONNECT"}
-
-	for _, op := range operations {
-		t.Run(op, func(t *testing.T) {
-			event := Event{
-				Object: obj,
-				Identifier: types.ResourceIdentifier{
-					Group:     "",
-					Version:   "v1",
-					Resource:  "testkinds",
-					Namespace: "test-ns",
-					Name:      "test-resource",
-				},
-				Operation: op,
-				UserInfo: UserInfo{
-					Username: "test-user",
-				},
-				Path: "",
-			}
-
-			message := GetCommitMessage(event)
-			expected := "[" + op + "] v1/testkinds/test-resource"
-			assert.Equal(t, expected, message)
-		})
-	}
 }
 
 func TestPathGeneration_ConsistentOutput(t *testing.T) {
@@ -545,92 +310,6 @@ func TestPathGeneration_ConsistentOutput(t *testing.T) {
 	assert.Equal(t, "v1/pods/consistent-ns/consistent-test.yaml", path1)
 }
 
-func TestGenerateLocalCommits_DeleteOperation(t *testing.T) {
-	// Test DELETE operation logic (file removal)
-	obj := &unstructured.Unstructured{}
-	obj.SetName("test-configmap")
-	obj.SetNamespace("default")
-	obj.SetKind("ConfigMap")
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "configmaps",
-			Namespace: "default",
-			Name:      "test-configmap",
-		},
-		Operation: "DELETE",
-		UserInfo: UserInfo{
-			Username: "admin",
-		},
-		Path: "",
-	}
-
-	// Verify commit message includes DELETE
-	commitMessage := GetCommitMessage(event)
-	assert.Contains(t, commitMessage, "[DELETE]")
-	assert.Contains(t, commitMessage, "configmaps/test-configmap")
-}
-
-func TestGenerateLocalCommits_CreateUpdateDeleteMixed(t *testing.T) {
-	// Test that different operations generate appropriate commit messages
-	testCases := []struct {
-		name      string
-		operation string
-		objName   string
-		expected  string
-	}{
-		{
-			name:      "CREATE operation",
-			operation: "CREATE",
-			objName:   "new-pod",
-			expected:  "[CREATE] v1/pods/new-pod",
-		},
-		{
-			name:      "UPDATE operation",
-			operation: "UPDATE",
-			objName:   "existing-pod",
-			expected:  "[UPDATE] v1/pods/existing-pod",
-		},
-		{
-			name:      "DELETE operation",
-			operation: "DELETE",
-			objName:   "old-pod",
-			expected:  "[DELETE] v1/pods/old-pod",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			obj := &unstructured.Unstructured{}
-			obj.SetName(tc.objName)
-			obj.SetNamespace("default")
-			obj.SetKind("Pod")
-
-			event := Event{
-				Object: obj,
-				Identifier: types.ResourceIdentifier{
-					Group:     "",
-					Version:   "v1",
-					Resource:  "pods",
-					Namespace: "default",
-					Name:      tc.objName,
-				},
-				Operation: tc.operation,
-				UserInfo: UserInfo{
-					Username: "test-user",
-				},
-				Path: "",
-			}
-
-			message := GetCommitMessage(event)
-			assert.Contains(t, message, tc.expected)
-		})
-	}
-}
-
 func TestToGitPath_DeleteOperation(t *testing.T) {
 	// Test that file paths are consistent regardless of operation
 	identifier := types.ResourceIdentifier{
@@ -646,150 +325,4 @@ func TestToGitPath_DeleteOperation(t *testing.T) {
 	expected := "v1/secrets/production/test-resource.yaml"
 
 	assert.Equal(t, expected, path)
-}
-
-func TestDeleteOperation_CommitMessageFormat(t *testing.T) {
-	// Test that DELETE operations have proper commit message format
-	testCases := []struct {
-		name      string
-		namespace string
-		resource  string
-		username  string
-		expected  string
-	}{
-		{
-			name:      "app-config",
-			namespace: "staging",
-			resource:  "configmaps",
-			username:  "developer",
-			expected:  "[DELETE] v1/configmaps/app-config",
-		},
-		{
-			name:      "db-secret",
-			namespace: "production",
-			resource:  "secrets",
-			username:  "admin",
-			expected:  "[DELETE] v1/secrets/db-secret",
-		},
-		{
-			name:      "web-deployment",
-			namespace: "default",
-			resource:  "deployments",
-			username:  "system:serviceaccount:kube-system:deployment-controller",
-			expected:  "[DELETE] apps/v1/deployments/web-deployment",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			obj := &unstructured.Unstructured{}
-			obj.SetName(tc.name)
-			obj.SetNamespace(tc.namespace)
-
-			group := ""
-			if tc.resource == "deployments" {
-				group = "apps"
-			}
-
-			event := Event{
-				Object: obj,
-				Identifier: types.ResourceIdentifier{
-					Group:     group,
-					Version:   "v1",
-					Resource:  tc.resource,
-					Namespace: tc.namespace,
-					Name:      tc.name,
-				},
-				Operation: "DELETE",
-				UserInfo: UserInfo{
-					Username: tc.username,
-				},
-				Path: "",
-			}
-
-			message := GetCommitMessage(event)
-			assert.Equal(t, tc.expected, message)
-		})
-	}
-}
-
-func TestDeleteOperation_ClusterScoped(t *testing.T) {
-	// Test DELETE operation for cluster-scoped resources
-	obj := &unstructured.Unstructured{}
-	obj.SetName("test-namespace")
-	obj.SetKind("Namespace")
-
-	identifier := types.ResourceIdentifier{
-		Group:     "",
-		Version:   "v1",
-		Resource:  "namespaces",
-		Namespace: "",
-		Name:      "test-namespace",
-	}
-
-	event := Event{
-		Object:     obj,
-		Identifier: identifier,
-		Operation:  "DELETE",
-		UserInfo: UserInfo{
-			Username: "cluster-admin",
-		},
-		Path: "cluster-repo",
-	}
-
-	// Verify file path
-	filePath := identifier.ToGitPath()
-	assert.Equal(t, "v1/namespaces/test-namespace.yaml", filePath)
-
-	// Verify commit message
-	commitMessage := GetCommitMessage(event)
-	assert.Equal(t, "[DELETE] v1/namespaces/test-namespace", commitMessage)
-}
-
-func TestBatchOperations_MultipleDeletes(t *testing.T) {
-	// Test that multiple DELETE operations can be processed
-	resources := []struct {
-		name      string
-		namespace string
-		plural    string
-	}{
-		{"pod-1", "default", "pods"},
-		{"pod-2", "default", "pods"},
-		{"service-1", "default", "services"},
-		{"configmap-1", "kube-system", "configmaps"},
-	}
-
-	var events []Event
-	for _, res := range resources {
-		obj := &unstructured.Unstructured{}
-		obj.SetName(res.name)
-		obj.SetNamespace(res.namespace)
-
-		event := Event{
-			Object: obj,
-			Identifier: types.ResourceIdentifier{
-				Group:     "",
-				Version:   "v1",
-				Resource:  res.plural,
-				Namespace: res.namespace,
-				Name:      res.name,
-			},
-			Operation: "DELETE",
-			UserInfo: UserInfo{
-				Username: "batch-delete-user",
-			},
-			Path: "",
-		}
-		events = append(events, event)
-	}
-
-	// Verify each event has correct DELETE operation
-	for i, event := range events {
-		message := GetCommitMessage(event)
-		assert.Contains(t, message, "[DELETE]")
-		assert.Contains(t, message, resources[i].name)
-	}
-
-	// Verify we have the expected number of events
-	assert.Len(t, events, 4)
 }
